+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Allow Symfony v7 dependencies.
 ### Fixed
 - Additional whitespace being output in PHP v8.3+ for header values which use
   encoded-words split across multiple lines.